staging: sync: Use an on-stack allocation for fence info ioctl
Since the fence info ioctl limits output data length to 4096 bytes, we
can just use a 4096-byte on-stack buffer for it. We have 16 KiB of stack
space, and this ioctl does not have deep function call depths, so using
4 KiB of the stack for a temporary buffer here is safe. This ioctl is
used for every frame rendered to the display, so eliminating dynamic
memory allocation overhead here improves performance.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: UtsavisGreat <utsavbalar1231@gmail.com>